refactor: infrastructure が application port を直接実装し Composition Root を純粋な組み立てに限定
GhClient に impl PromptStatusPort、TomlConfigRepository に impl LoadConfigPort /
UpdateConfigPort を追加し、bin 層にあった転送のみのアダプタ
(GhPromptAdapter, TomlConfigPortAdapter) を削除した。

クロスコンテキスト型である ConfigAdapter は src/adapters.rs に切り出し、
app.rs をインスタンス生成と配線のみを行う Composition Root に整理した。
