+++ src/adapters.rs
@@ -1,0 +1,34 @@
+use crate::contexts::config::application::config_service::ConfigService;
+use crate::contexts::config::infrastructure::toml_loader::TomlConfigRepository;
+use crate::contexts::prompt::application::{OutputToken, errors::ErrorToken};
+use crate::contexts::prompt::interface::presentation::PresentationConfigPort;
+
+pub(crate) struct ConfigAdapter(ConfigService);
+
+impl ConfigAdapter {
+    pub(crate) fn load() -> Self {
+        Self(ConfigService::new(&TomlConfigRepository))
+    }
+}
+
+impl PresentationConfigPort for ConfigAdapter {
+    fn render_token(&self, token: &OutputToken) -> String {
+        match token {
+            OutputToken::MergeReady => self.0.render_merge_ready(),
+            OutputToken::Conflict => self.0.render_conflict(),
+            OutputToken::UpdateBranch => self.0.render_update_branch(),
+            OutputToken::SyncUnknown => self.0.render_sync_unknown(),
+            OutputToken::CiFail => self.0.render_ci_fail(),
+            OutputToken::CiAction => self.0.render_ci_action(),
+            OutputToken::ReviewRequested => self.0.render_review(),
+        }
+    }
+
+    fn render_error_token(&self, token: ErrorToken) -> String {
+        match token {
+            ErrorToken::AuthRequired => self.0.render_auth_required(),
+            ErrorToken::RateLimited => self.0.render_rate_limited(),
+            ErrorToken::ApiError => self.0.render_api_error(),
+        }
+    }
+}
