+++ src/contexts/prompt/infrastructure/gh.rs
@@ -5,6 +5,7 @@ use std::time::{Duration, Instant};
 
 use serde::Deserialize;
 
+use crate::contexts::prompt::application::port::PromptStatusPort;
 use crate::contexts::prompt::domain::branch_sync::{BranchSync, BranchSyncStatus};
 use crate::contexts::prompt::domain::ci_checks::{CheckBucket, CiChecks};
 use crate::contexts::prompt::domain::error::RepositoryError;
@@ -102,26 +103,20 @@ impl GhClient {
     }
 }
 
-impl GhClient {
-    /// # Errors
-    /// Returns `RepositoryError` if the PR lifecycle cannot be fetched.
-    pub fn fetch_lifecycle(&self) -> Result<PrLifecycle, RepositoryError> {
+impl PromptStatusPort for GhClient {
+    fn fetch_lifecycle(&self) -> Result<PrLifecycle, RepositoryError> {
         let raw = self.pr_view_cached()?;
         Ok(translate_lifecycle(&raw.state))
     }
 
-    /// # Errors
-    /// Returns `RepositoryError` if the sync status cannot be fetched.
-    pub fn fetch_sync_status(&self) -> Result<BranchSync, RepositoryError> {
+    fn fetch_sync_status(&self) -> Result<BranchSync, RepositoryError> {
         let raw = self.pr_view_cached()?;
         let behind_by =
             fetch_behind_by(&raw.base_ref_name, &raw.head_ref_name, self.cwd.as_deref());
         Ok(BranchSync::new(translate_sync(&raw.mergeable, behind_by)))
     }
 
-    /// # Errors
-    /// Returns `RepositoryError` if the CI checks cannot be fetched.
-    pub fn fetch_checks(&self) -> Result<CiChecks, RepositoryError> {
+    fn fetch_checks(&self) -> Result<CiChecks, RepositoryError> {
         let bytes = match self.run_gh(&["pr", "checks", "--json", "bucket,state"]) {
             Ok(b) => b,
             // CI が未設定のブランチではチェックなし（正常）として扱う
@@ -137,18 +132,14 @@ impl GhClient {
         ))
     }
 
-    /// # Errors
-    /// Returns `RepositoryError` if the review state cannot be fetched.
-    pub fn fetch_review(&self) -> Result<Review, RepositoryError> {
+    fn fetch_review(&self) -> Result<Review, RepositoryError> {
         let raw = self.pr_view_cached()?;
         Ok(Review::new(translate_review(
             raw.review_decision.as_deref(),
         )))
     }
 
-    /// # Errors
-    /// Returns `RepositoryError` if the merge readiness cannot be fetched.
-    pub fn fetch_readiness(&self) -> Result<MergeReadiness, RepositoryError> {
+    fn fetch_readiness(&self) -> Result<MergeReadiness, RepositoryError> {
         let raw = self.pr_view_cached()?;
         Ok(translate_merge_readiness(
             raw.is_draft,
