+++ src/main.rs
@@ -1,3 +1,4 @@
+mod adapters;
 mod app;
 mod cli;
 mod contexts;
